Corregir errores y subir versión local

+++ Backend/quickbite-menu-service/src/main/java/com/quickbite/menu_service/entity/MenuItemIngredient.java
@@ -30,7 +30,7 @@ public class MenuItemIngredient {
     private Long ingredientId;
 
     @Column(name = "quantity", nullable = false)
-    private Double quantity;
+    private Integer quantity;
 
     @Column(name = "unit", nullable = false)
     private String unit;
